+++ src/ts/uimanager.ts
@@ -27,14 +27,12 @@ export interface InternalUIConfig extends UIConfig {
 export interface UIConditionContext {
   /**
    * Tells if the player is loading or playing an ad.
-   * @deprecated to be updated for the upcoming native ads player v8 module
    */
   isAd: boolean;
   /**
-   * Tells the ad client (e.g. 'vast, 'ima') if {@link #isAd} is true.
-   * @deprecated to be updated for the upcoming native ads player v8 module
+   * Tells if the current ad requires an external UI, if {@link #isAd} is true.
    */
-  adClientType: string;
+  adRequiresUi: boolean;
   /**
    * Tells if the player is currently in fullscreen mode.
    */
@@ -226,7 +224,7 @@ export class UIManager {
         switch (event.type) {
           // When the ad starts, we store the event data
           case player.exports.PlayerEvent.AdStarted:
-            adStartedEvent = <AdEvent>event;
+            adStartedEvent = event as AdEvent;
             break;
           // When the ad ends, we delete the event data
           case player.exports.PlayerEvent.AdFinished:
@@ -244,11 +242,17 @@ export class UIManager {
       }
 
       // Detect if an ad has started
-      let ad = adStartedEvent != null;
+      let isAd = adStartedEvent != null;
+      let adRequiresUi = isAd;
+      if (isAd) {
+        let ad = adStartedEvent.ad;
+        // for now preventing showing an ad if it's non linear but requesting an UI
+        adRequiresUi = ad.isLinear && ad.requiresUi;
+      }
 
       this.resolveUiVariant({
-        isAd: ad,
-        adClientType: null,
+        isAd: isAd,
+        adRequiresUi: adRequiresUi,
       }, (context) => {
         // If this is an ad UI, we need to relay the saved ON_AD_STARTED event data so ad components can configure
         // themselves for the current ad.
@@ -351,7 +355,7 @@ export class UIManager {
     // Determine the current context for which the UI variant will be resolved
     const defaultContext: UIConditionContext = {
       isAd: false,
-      adClientType: null,
+      adRequiresUi: false,
       isFullscreen: this.player.getViewMode() === this.player.exports.ViewMode.Fullscreen,
       isMobile: BrowserUtils.isMobile,
       isPlaying: this.player.isPlaying(),